+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2025-09-06
+### Added
+- New command with gates annotate
+- VEP annotation
+- Formatting for VCF into TSV
+  
+### Changed
+- Log file name so that it includes command
+- Automatic sample name detection for gates call
+
 ## [1.1.0] - 2025-09-06
 ### Added
 - Support for germline variant calling using HaplotypeCaller.